https://www.youtube.com/watch?v=EswTAF1pau4
Yeah it's a video link, FRUSTRATION!!! Gradle can't handle spaces in
some lines like for instance if you are doing

replace "@VERSION@", project.version

that little space right after the comma makes it error out and not
replace the symbol. Remove that little space and everything is fine. And
yet, it knows enough that the commands have spaces after them!

Oversight! I need to report this somewhere as it can't be a design
feature!