1.12.2-2.0.4
Added Translations:

Russian

Czech

German (DE)

Added more of the in-mod phrasing to the translation tables.

Added 16 textures for allowing you to adjust the torch placement indicator according to the magnitude of a light source. It does not autodetect. But you can cycle it with the comma and period keys.

Added enableOptimize and optimizedPlacement to the clickable gui. The enableOptimize attribute is automatically turned on when you use the cycling keys.